feat(ui): add sliver scaffold and reorganize ui components
- Implement ScaffoldSliver for sliver-based layouts
- Create CustomSliverAppBar with theme and language controls
- Reorganize ui components into proper directory structure
- Add ThemeToggleButton and LanguageSelector widgets
